Fixes #61: moved UI classes to the UI package
