+++ app/src/main/java/com/tiptopgoodstudio/androidresources/ui/SettingsFragment.java
@@ -1,12 +1,13 @@
-package com.tiptopgoodstudio.androidresources;
+package com.tiptopgoodstudio.androidresources.ui;
 
-import android.content.Context;
-import android.net.Uri;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.tiptopgoodstudio.androidresources.R;
+import com.tiptopgoodstudio.androidresources.ui.HomeFragment;
 
 public class SettingsFragment extends Fragment {
