+++ app/src/main/res/layout-land/fragment_home.xml
@@ -6,7 +6,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.tiptopgoodstudio.androidresources.HomeFragment">
+    tools:context="com.tiptopgoodstudio.androidresources.ui.HomeFragment">
 
     <android.support.constraint.Guideline
         android:id="@+id/guidelineH"
